requirements.txt
# Core dependencies
# vercel-blob==0.4.2 (replaced by local SDK)
fastapi==0.103.1
uvicorn==0.23.2
pydantic~=2.7.1
pydantic-settings==2.0.3
python-dotenv==1.0.0
typing-extensions~=4.10.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
# yfinance==0.2.33  # Disabled: Package too large for Vercel serverless function limit

# Database drivers and clients
motor>=3.3.2
pymongo>=4.6.1

# Extra utilities
toolz==0.12.0  # Functional programming helpers
loguru==0.7.2  # Better logging
tenacity==8.2.3  # Retry logic
jsonschema==4.19.0  # JSON schema validation
aiofiles==23.2.1  # Async file operations
email-validator==2.0.0  # For email validation in Pydantic
psutil==7.0.0  # For system metrics in health checks
httpx==0.27.0 # For making HTTP requests, required by some automation handlers
requests>=2.31.0 # Required by local Vercel Blob SDK
tqdm==4.66.1 # Required by local Vercel Blob SDK for progress bars
aiohttp>=3.8.0

# Specify urllib3 version that works with botocore
urllib3<1.27,>=1.25.4
# Use vercel-blob version that's compatible with the requests version
vercel-blob>=0.3.0  # Vercel Blob Storage for serverless file storage
